pre & post treat; calc loss

+++ src/ELBO.cpp
@@ -6,7 +6,7 @@ arma::vec elementwise_prod(const arma::vec & yv,
                           const arma::uvec & yi,
                           const arma::vec & R){
   arma::vec out = arma::zeros<arma::vec>(R.n_rows);
-  for(int i=0; i<yv.n_rows; i++){
+  for(int i=0; i < (int) yv.n_rows; i++){
     out(yi(i)) = yv(i)*R(yi(i));
   }
   return out;
